merge

+++ ARMS-Backend/src/main/java/ca/mcgill/ecse321/arms/dao/BusinessHourRepository.java
@@ -5,6 +5,8 @@ import org.springframework.data.repository.CrudRepository;
 import ca.mcgill.ecse321.arms.model.BusinessHour;
 
 public interface BusinessHourRepository extends CrudRepository<BusinessHour, String> {
-    BusinessHour findBusinessHourByBusinessHourID(int id);
-    Integer deleteBusinessHourByBusinessHourID(int id);
+
+    BusinessHour findBusinessHourByBusinessHourID(long id);
+    Integer deleteBusinessHourByBusinessHourID(long id);
+
 }
